Issues
The problem is in the DialogueManager script. DisplayDialogue or UpdateDialogue functions are instantiating the scriptable object lines over and over again at what seems to me like random points. I don't understand what changed so I don't even know where to start fixing it. I added a for loop on ln 156 so that I could actually use the spacebar to move my dialogue along, but I can only assume the problem is with that. Still, when I took out the for loop and tried to re-use the on-screen continue button the dialogue was still messed up and repeating itself. It either doesn't cycle through the dialogue at all or randomly repeats itself.